requirements.txt
numpy==1.19.5
opencv-contrib-python-headless==4.5.1.48
paho-mqtt==1.5.1
